requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlmodel==0.0.14
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
pydantic==2.5.0